fix bugs

+++ src/application/use-cases/calendar/delete-calendar-event.use-case.ts
@@ -13,5 +13,7 @@ export const deleteCalendarEventUseCase = (
     deleteNotificationChannelUseCase: IDeleteNotificationChannelUseCase
 ) => async (opts: DeleteCalendarEventOptions) => {
     await calendarService.deleteCalendarEvent(opts);
-    await deleteNotificationChannelUseCase(opts.eventId);
+    if (process.env.NODE_ENV === "production" && opts.channel) {
+        await deleteNotificationChannelUseCase(opts.eventId);
+    }
 }
